+++ HWM.CheckoutApp.BusinessObjects/Model/Order.cs
@@ -11,15 +11,15 @@ namespace HWM.CheckoutApp.Model
         public int OrderID { get; set; }
         public byte PaymentMethodTypeID { get; set; }
         public System.DateTime OrderDate { get; set; }
-        public decimal TotalAmountPaid { get; set; }
+        public double TotalAmountPaid { get; set; }
 
 
 
-        public decimal? Discount { get; set; }
+        public double? Discount { get; set; }
         public int PaymentTransactionNumber { get; set; }
         public string OrderTrackingNumber { get; set; }
-        public decimal? VATCharge { get; set; }
-        public decimal TotalCost { get; set; }
+        public double? VATCharge { get; set; }
+        public double TotalCost { get; set; }
         public System.DateTime PaymentTransactionDate { get; set; }
